Fix a flaw in the data model.
Using fine-grained timestamps with timezones for reservations is very problematic, I've switched the reservation to use Date values instead of timestamps with timezones.

Date values side-step a huge problem related to search and creating reservations. Once a time interval is introduced, performing a search needs to resolve the timezone of the location being searched and -- if we are going for completeness -- must account for the check-in and check-out time of each individual unit... yikes.

Normally, I'd introduce another migration, but because the system should never be deployed as was done previously, I just updated the previous migration. Somewhat bad form, but I think it's ok in this context.