Add price to unit.
This is a very (very, very, very) naive way of doing pricing. It's also somewhat wrong: people that reserve a unit before a price change shouldn't be affected by the price change... or should they? Is it fair to charge people more when the price goes down? It doesn't seem fair to charge them more if the price goes up...

I've noticed some cruft while making this change:

1. Flash messages aren't assigned to keys with conventional names; they should probably either use the key :notice or :alert.
2. Factory data used by cucumber is specified by the factories in the test directory. Duplicating this code doesn't seem ideal... but it could lead to some confusion.